+++ src/Collision/Collision.h
@@ -11,26 +11,26 @@ namespace collision{
     // 矩形の当たり判定
     struct Rectangle{
         game::Game* game; // ゲームオブジェクト
-        int16_t right_display_x; // 右上の画面x座標
-        int16_t top_display_y; // 左上の画面y座標
-        int16_t left_display_x; // 左下の画面x座標
-        int16_t bottom_display_y; // 右下の画面y座標
+        double right_display_x; // 右上の画面x座標
+        double top_display_y; // 左上の画面y座標
+        double left_display_x; // 左下の画面x座標
+        double bottom_display_y; // 右下の画面y座標
 
         // コンストラクタ
         Rectangle();
-        Rectangle(game::Game* game, int16_t right_display_x, int16_t top_display_y, int16_t left_display_x, int16_t bottom_display_y);
+        Rectangle(game::Game* game, double right_display_x, double top_display_y, double left_display_x, double bottom_display_y);
     };
 
     // 円形の当たり判定
     struct Circle{
         game::Game* game; // ゲームオブジェクト
-        int16_t center_display_x; // 中心の画面x座標
-        int16_t center_display_y; // 中心の画面y座標
-        uint16_t radius; // 半径
+        double center_display_x; // 中心の画面x座標
+        double center_display_y; // 中心の画面y座標
+        double radius; // 半径
 
         // コンストラクタ
         Circle();
-        Circle(game::Game* game, int16_t center_display_x, int16_t center_display_y, uint16_t radius);
+        Circle(game::Game* game, double center_display_x, double center_display_y, double radius);
     };
 
     // 矩形同士の当たり判定
